+++ components/Character/CarouselDisplay/index.tsx
@@ -1,13 +1,8 @@
 import React, { ReactNode, FC } from 'react'
-import Image, { ImageLoader } from 'next/image'
-import Button from 'components/Custom/Button/index'
 import type { Character } from 'lib/types'
 interface CarouselProps {
   children?: ReactNode
   chr?: Character[]
-}
-interface CarouselItemProps {
-  chr: Character
 }
 
 /** The single-page view of character images and buttons which navigate to individual character pages. */
@@ -24,27 +19,6 @@ const CarouselDisplay: FC = (props: CarouselProps) => {
             </div>
           )
         })}
-      </div>
-    </div>
-  )
-}
-
-/** A single character image and its button in the carousel. */
-export const CarouselItem = ({ chr }: CarouselItemProps): JSX.Element => {
-  return (
-    <div className='flex flex-col items-center'>
-      {chr.image && (
-        <div>
-          <Image
-            src={chr.image}
-            width={400}
-            height={300}
-            alt='Picture of the author'
-          ></Image>
-        </div>
-      )}
-      <div className='my-12'>
-        <Button>{chr.name}</Button>
       </div>
     </div>
   )
